Makes some installations conditional to reduce workload

+++ .buildkite/install.sh
@@ -5,9 +5,10 @@ set -eu
 echo "###############################"
 echo "# Installing APT Repositories #"
 echo "###############################"
+release_name=$(lsb_release -c)
 sudo apt-add-repository -y ppa:ubuntu-toolchain-r/test
-wget -O - http://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
-sudo apt-add-repository -y "deb http://apt.llvm.org/$(lsb_release -c)/ llvm-toolchain-$(lsb_release -c)-6.0 main"
+wget -O - "http://apt.llvm.org/llvm-snapshot.gpg.key" | sudo apt-key add -
+sudo apt-add-repository -y "deb http://apt.llvm.org/${release_name}/ llvm-toolchain-${release_name}-6.0 main"
 sudo apt update
 
 echo "###############################"
@@ -77,30 +78,53 @@ sudo /usr/bin/update-alternatives --remove-all gcc \
                                          --slave /usr/bin/g++ g++ /usr/bin/g++-7 \
                                          --slave /usr/bin/gfortran gfortran /usr/bin/gfortran-7
 
-mkdir -p artifacts
-cd artifacts
+mkdir -p install_artifacts && cd install_artifacts
 
 echo "###############################"
-echo "#   Installing CMake v3.5.1   #"
+echo "#      Installing CMake       #"
 echo "###############################"
-/usr/bin/wget https://cmake.org/files/v3.5/cmake-3.5.1-Linux-x86_64.sh
-/bin/sh cmake-3.5.1-Linux-x86_64.sh --prefix=/usr/local --exclude-subdir
-sudo cp puppet/modules/files/files/FindBoost.cmake /usr/local/share/cmake-3.5/Modules/FindBoost.cmake
+cmake_installed_version=$(cmake --version 2>&1 | head -n1 | cut -d' ' -f3)
+
+# Only download cmake if it either isn't installed already
+if [ "x${cmake_installed_version}" != "x${CMAKE_VERSION}" ]
+then
+    major_version=$(echo "${CMAKE_VERSION}" | cut -d'.' -f1,2)
+    /usr/bin/wget "https://cmake.org/files/v${major_version}/cmake-${CMAKE_VERSION}-Linux-x86_64.sh"
+    /bin/sh "cmake-${CMAKE_VERSION}-Linux-x86_64.sh" --prefix=/usr/local --exclude-subdir
+    sudo cp "puppet/modules/files/files/FindBoost.cmake" "/usr/local/share/cmake-${major_version}/Modules/FindBoost.cmake"
+fi
 
 echo "###############################"
 echo "#   Installing Intel OpenCL   #"
 echo "###############################"
-mkdir intel-opencl
-cd intel-opencl
-wget http://registrationcenter-download.intel.com/akdlm/irc_nas/11396/SRB5.0_linux64.zip
-unzip SRB5.0_linux64.zip
-mkdir root
-for i in *.tar.xz; do tar -C root -xf "$i"; done
-sudo cp -r root/* /
-sudo ldconfig
+if [ ! -e "/opt/intel/opencl/libOpenCL.so" ]
+then
+    mkdir intel-opencl
+    cd intel-opencl
+    wget http://registrationcenter-download.intel.com/akdlm/irc_nas/11396/SRB5.0_linux64.zip
+    unzip SRB5.0_linux64.zip
+    mkdir root
+    for i in *.tar.xz; do tar -C root -xf "$i"; done
+    sudo cp -r root/* /
+    sudo ldconfig
+fi
 
 echo "###############################"
 echo "#     Installing Toolchain    #"
 echo "###############################"
-wget -N http://nubots.net/debs/nubots-toolchain-3.0.1-travis.deb
-sudo dpkg -i nubots-toolchain-3.0.1-travis.deb
+toolchain_installed_version=$(dpkg -l | grep '^ii' | grep nubots-toolchain | awk '{print $3}')
+
+# Only download the toolchain if it either isn't installed already or we want a different version
+if [ "x${toolchain_installed_version}" != "x${TOOLCHAIN_VERSION}" ]
+then
+    # Remove the toolchain if the wrong version is installed
+    if [ "x${toolchain_installed_version}" != "x" ]
+    then
+        sudo dpkg -P nubots-toolchain
+    fi
+
+    # Download and install the correct version
+    wget -N "http://nubots.net/debs/nubots-toolchain-${TOOLCHAIN_VERSION}.deb"
+    sudo dpkg -i "nubots-toolchain-${TOOLCHAIN_VERSION}.deb"
+fi
+
